graphql_vehicles/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
strawberry-graphql[fastapi]==0.240.0
pydantic==2.8.2
packaging>=24.0